Made solution look more like in-class
