+++ db/connection.rb
@@ -1,0 +1,4 @@
+ActiveRecord::Base.establish_connection(
+  adapter: "postgresql",
+  database: "moma_db"
+)
