+++ models/artist.rb
@@ -1,0 +1,3 @@
+class Artist < ActiveRecord::Base
+  has_many :paintings
+end
