+++ models/painting.rb
@@ -1,0 +1,3 @@
+class Painting < ActiveRecord::Base
+  belongs_to :artist
+end
