+++ moma.rb
@@ -1,21 +1,10 @@
 require 'active_record'
 require 'pry'
 
-ActiveRecord::Base.establish_connection(
-  adapter: "postgresql",
-  host: "localhost",
-  username: "peterlai",
-  password: "",
-  database: "moma_db"
-)
-
-class Artist < ActiveRecord::Base
-  has_many :paintings
-end
-
-class Painting < ActiveRecord::Base
-  belongs_to :artist
-end
+require_relative "db/connection"
+require_relative "models/artist"
+require_relative "models/painting"
 
 binding.pry
 
+"pry"
